Move duplication in the localization tests.
